requirements.txt
tldextract
requests
dnspython
colorama
python-whois
